requirements.txt
Django==4.2.28
django-jazzmin==3.0.2
python-dotenv==1.2.1
whitenoise==6.11.0
gunicorn==25.0.3
dj-database-url==3.1.0
psycopg2-binary==2.9.11
